funasr/requirements.txt
fastapi==0.103.0
jsonpatch==1.32
jsonpointer==2.1
typing_extensions==4.9.0
uvicorn==0.27.1
websocket-client==1.7.0
websockets==10.4